% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{aggregate_phydo_traits}
\alias{aggregate_phydo_traits}
\title{Aggregate phydo traits}
\usage{
aggregate_phydo_traits(location_realm_biome, eol_tbl)
}
\arguments{
\item{location_realm_biome}{data frame with locations}

\item{eol_tbl}{A data frame with species in the first column and traits in the rest}
}
\value{
A list with two data frames: one with qualitative traits and one with numeric traits
}
\description{
Aggregate phydo traits
}
