% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{phydo_convert_to_Beaulieu_data}
\alias{phydo_convert_to_Beaulieu_data}
\title{Convert the phydo format to one usable for Beaulieu-style}
\usage{
phydo_convert_to_Beaulieu_data(phydo)
}
\arguments{
\item{phydo}{The phydo object}
}
\value{
a list of two objects, phy and data, with the data in the format preferred by Beaulieu
}
\description{
Many of the packages authored by Jeremy Beaulieu with collaborators have the first column used to hold taxon names rather than store them as row names. Bless his heart. This converts a phydo object, which has its data with taxon names as row names, into having a data object internally with a column for taxon names.
}
