% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{construct_phydo_targets}
\alias{construct_phydo_targets}
\title{Do phydo analysis with targets}
\usage{
construct_phydo_targets(taxon, ncores = 1)
}
\arguments{
\item{taxon}{Clade of interest}

\item{ncores}{Number of cores to use}
}
\value{
Nothing, though a file is created in the current working directory
}
\description{
Do phydo analysis with targets
}
