% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoldata.R
\name{eol_traits2}
\alias{eol_traits2}
\title{Collapse the data frame produced by eol_data to a single row with traits as columns}
\usage{
eol_traits2(eol_df)
}
\arguments{
\item{eol_df}{A data frame produced by the function eol_data}
}
\value{
A data frame of one row with columns: behavior circadian rhythm, developmental mode, visual system, and wing morphology
}
\description{
Collapse the data frame produced by eol_data to a single row with traits as columns
}
