% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_combie_image_summary}
\alias{get_combie_image_summary}
\title{Get Wikipedia images}
\usage{
get_combie_image_summary(taxon)
}
\arguments{
\item{taxon}{}
}
\value{
images from Wikipedia
}
\description{
Get Wikipedia images
}
