% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_datelife_all}
\alias{get_datelife_all}
\title{Get all tree from datelife}
\usage{
get_datelife_all(taxon)
}
\arguments{
\item{taxon}{Clade of interest}
}
\value{
phylo object
}
\description{
Get all tree from datelife
}
