% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_descendant_species}
\alias{get_descendant_species}
\title{Get all descendant species of the taxon}
\usage{
get_descendant_species(taxon, strict = TRUE)
}
\arguments{
\item{taxon}{Clade of interest}
}
\value{
vector of species names
}
\description{
Uses taxize, datelife, and sources of GBIF and OpenTree
}
