% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_location_realm_biome}
\alias{get_location_realm_biome}
\title{Get location, realm, and biome}
\usage{
get_location_realm_biome(gbif_record)
}
\arguments{
\item{gbif_record}{Data from GBIF}
}
\value{
list containing a data.frame of species and locations, a table of realms (biogeographic regions), and a table of biomes
}
\description{
Get location, realm, and biome
}
