% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{get_phydo_data}
\alias{get_phydo_data}
\title{Get phydo raw data}
\usage{
get_phydo_data(taxon)
}
\arguments{
\item{taxon}{Clade of interest}
}
\value{
A list with all output
}
\description{
Get phydo raw data
}
