% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_pubmed_count_only}
\alias{get_pubmed_count_only}
\title{Get information on the count of papers from pubmed}
\usage{
get_pubmed_count_only(taxon, search.string = " AND phylogeny")
}
\arguments{
\item{taxon}{The clade of interest}

\item{search.string}{Include spaces and AND and similar search elements}
}
\value{
List with a count element (integer) and a recent.papers element (data.frame)
}
\description{
This will give the total number of pubmed articles mentioning the taxon name and other search string and information on the most recent such papers.
}
\examples{
taxon <- "Formicidae"
results <- get_pubmed_count_only(taxon)
print(paste("There are", results, "papers on", taxon, "and phylogeny"))
}
