% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_pubmed_info_on_OTOL}
\alias{get_pubmed_info_on_OTOL}
\title{Aggregate pubmed data down OToLtree}
\usage{
get_pubmed_info_on_OTOL(taxon, ignore_subspecies = TRUE)
}
\arguments{
\item{taxon}{The clade to investigate}

\item{ignore_subspecies}{Should subspecies be ignored?}
}
\value{
List with tree and information about nodes and tips
}
\description{
Aggregate pubmed data down OToLtree
}
\examples{
taxon <- "Pinaceae"
info <- get_pubmed_info_on_OTOL(taxon)
}
