% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{get_wikipedia_pics}
\alias{get_wikipedia_pics}
\title{Get Wikipedia images}
\usage{
get_wikipedia_pics(taxon)
}
\arguments{
\item{taxon}{}
}
\value{
images from Wikipedia
}
\description{
Get Wikipedia images
}
