% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{phydo_drop_type}
\alias{phydo_drop_type}
\title{Return a phydo object with only the selected kind of traits}
\usage{
phydo_drop_type(phydo, keep = c("continuous", "discrete"))
}
\arguments{
\item{phydo}{a phydo class object, with tree and data}

\item{keep}{Whether you want to keep continuous or discrete data}
}
\value{
A phydo object with the tree and only the chosen kind of data
}
\description{
Return a phydo object with only the selected kind of traits
}
