% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{plot_gis}
\alias{plot_gis}
\title{Plot a map of points}
\usage{
plot_gis(gbif_points)
}
\arguments{
\item{gbif_points}{Data.frame containing points (latitude and longitude, perhaps other data as columns)}
}
\value{
A ggplot2 map

Note that we will need to set zoom automatically, also add points, also make the ocean full of water.
}
\description{
Plot a map of points
}
\examples{
gis_points <- spocc_taxon_query("Tyto", limit=100, sources="gbif")
plot_gis(gbif_points)
}
