% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{plot_pubmed_info_on_OTOL}
\alias{plot_pubmed_info_on_OTOL}
\title{Plot pubmed data down OToLtree}
\usage{
plot_pubmed_info_on_OTOL(pubmed_info_on_OTOL)
}
\arguments{
\item{pubmed_info_on_OTOL}{List with tree and information about nodes and tips, from get_pubmed_info_on_OTOL}
}
\value{
Plot
}
\description{
This will use reconstruction of number of papers down a tree
}
