% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{reconstruct_apomorphy}
\alias{reconstruct_apomorphy}
\title{Reconstruct an apomorphy down a tree using parsimony}
\usage{
reconstruct_apomorphy(phy, data, apomorphy)
}
\arguments{
\item{phy}{A phylo object}

\item{data}{A phyDat class object containing the states}

\item{apomorphy}{The state that is the apomorphy}
}
\value{
A tree with node.labels with the reconstructed states
}
\description{
Given a phylogeny, tip states, and which state is the apomorphy, reconstruct the apomorphy down the tree.
}
