% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{render_quarto}
\alias{render_quarto}
\title{Create a file of results}
\usage{
render_quarto(phydo_result, taxon)
}
\arguments{
\item{phydo_result}{The result of run_phydo}

\item{taxon}{Clade of interest}
}
\value{
Nothing, though a file is created in the current working directory
}
\description{
Create a file of results
}
