% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{reroot.tree}
\alias{reroot.tree}
\title{Root a tree}
\usage{
reroot.tree(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
A rooted tree
}
\description{
From Jeremy Beaulieu, slight mods from Brian O'Meara
}
