% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{run_corHMM}
\alias{run_corHMM}
\title{Run a corHMM analysis}
\usage{
run_corHMM(phy, trait_data, ncores)
}
\arguments{
\item{phy}{A phylo object}

\item{trait_data}{A data.frame with species in the first column and traits in the rest}

\item{ncores}{Number of cores to use}
}
\value{
A list with the results of the corHMM analysis
}
\description{
Run a corHMM analysis
}
