% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runanalysis.R
\name{run_misse}
\alias{run_misse}
\title{Run misse analysis}
\usage{
run_misse(phy, species = NULL, ncores = 1)
}
\arguments{
\item{phy}{The phylo object}

\item{species}{A vector of species names in the clade, NOT just on the tree.}
}
\value{
A list with class misse.fits objects
}
\description{
Run a misse analysis to estimate parameters of diversification models.
By passing in a vector of species names, it uses those to estimate the
sampling fraction based on how many are in the tree.
}
