% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{WWFload}
\alias{WWFload}
\title{Use azizka/speciesgeocodeR/ and WWF data to encode locations for habitat and biome}
\usage{
WWFload(x = NULL)
}
\arguments{
\item{x}{Data.frame containing points (latitude and longitude, perhaps other data as columns)}
}
\value{
data.frame with columns for habitat and biome.
}
\description{
Use azizka/speciesgeocodeR/ and WWF data to encode locations for habitat and biome
}
